problem with fire
