+++ 02-model.py
@@ -8,16 +8,15 @@ from sklearn.pipeline import make_pipeline
 from sklearn.preprocessing import LabelBinarizer, StandardScaler
 from sklearn_pandas import DataFrameMapper
 
-df = pd.read_csv("data/football.csv", parse_dates=[6])
+df = pd.read_csv("data/basketball.csv", parse_dates=[3])
 df = df.sort_values(["name", "date"]).reset_index(drop=True)
-df["yards"] = df["passing"] + df["rushing"] + df["receiving"]
-df["yards_1"] = df.groupby("name")["yards"].shift(1)
-df["yards_2"] = df.groupby("name")["yards"].shift(2)
-df = df.dropna(subset=["yards_1", "yards_2"])
+df["points_1"] = df.groupby("name")["points"].shift(1)
+df["points_2"] = df.groupby("name")["points"].shift(2)
+df = df.dropna(subset=["points_1", "points_2"])
 
-target = "yards"
+target = "points"
 y = df[target]
-X = df[["position", "yards_1", "yards_2"]]
+X = df[["position", "points_1", "points_2"]]
 
 X_train, X_test, y_train, y_test = train_test_split(
     X, y, test_size=0.10, random_state=42, shuffle=False
@@ -26,8 +25,8 @@ X_train, X_test, y_train, y_test = train_test_split(
 mapper = DataFrameMapper(
     [
         (["position"], [SimpleImputer(strategy="most_frequent"), LabelBinarizer()]),
-        (["yards_1"], [SimpleImputer(), StandardScaler()]),
-        (["yards_2"], [SimpleImputer(), StandardScaler()]),
+        (["points_1"], [SimpleImputer(), StandardScaler()]),
+        (["points_2"], [SimpleImputer(), StandardScaler()]),
     ],
     df_out=True,
 )
