+++ 05-database.py
@@ -4,7 +4,7 @@ import sqlite3
 import pandas as pd
 from fire import Fire
 
-con = sqlite3.connect("data/football.db")
+con = sqlite3.connect("data/basketball.db")
 
 with open("pickles/pipe.pkl", "rb") as f:
     pipe = pickle.load(f)
@@ -14,8 +14,7 @@ def fetch_player_data(name):
     player = pd.read_sql(
         f"""
         select
-        *,
-        (passing + rushing + receiving) as yards
+        *
         from players
         where name = '{name}'
         order by date desc
@@ -30,8 +29,8 @@ def prep_data(player):
     X_new = pd.DataFrame(
         {
             "position": [player.position[0]],
-            "yards_1": [player.yards[0]],
-            "yards_2": [player.yards[1]],
+            "points_1": [player.points[0]],
+            "points_2": [player.points[1]],
         }
     )
     return X_new
@@ -41,13 +40,13 @@ def predict(player):
     X = fetch_player_data(player)
     date = X["date"].max()
     X = prep_data(X)
-    yards = round(pipe.predict(X)[0], 2)
+    points = round(pipe.predict(X)[0], 2)
     df = pd.DataFrame(
         {
             "date_created": pd.Timestamp("now"),
             "name": [player],
             "last_game": [date],
-            "yards": [yards],
+            "points": [points],
         }
     )
     df.to_sql("predictions", con, if_exists="append", index=False)
